Check for "all" and shorthand properties
When the CSS specifies "all" as the CSS property to transition we still get the individual CSS properties that have changed. If the caller specified "all" as the property to monitor we can just resolve the promise at the first time the transitionend event has been received.
If the CSS specifies "all" and the caller wants to monitor a specific CSS property, it should still work just fine. In this case we can just resolve the promise as soon as the event handler is called for the first time.
In addition we need to take into consideration if the property has its shorthand form defined as a property to transition. This can affect which duration we have to use.